add DTO controller, entity

+++ src/main/java/com/projetCertif/controllerDto/ChannelControllerDto.java
@@ -1,12 +1,14 @@
 package com.projetCertif.controllerDto;
 
 import com.projetCertif.dao.entity.Channel;
+import com.projetCertif.dao.entityDto.ChannelDto;
 import com.projetCertif.service.ChannelService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,40 +21,53 @@ public class ChannelControllerDto {
 
     // GET ALL CHANNELS
     @GetMapping("channels")
-    public ResponseEntity<List<Channel>> getAllChannels() {
-        return ResponseEntity.ok(channelService.getAllChannels());
+    public List<ChannelDto> getAllChannels(){
+        List<Channel> channels = channelService.getAllChannels();
+        List<ChannelDto> channelDtos = new ArrayList<>();
+        for(Channel c : channels) {
+            channelDtos.add(ChannelDto.fromEntity(Optional.ofNullable((c))));
+        }
+        return channelDtos;
+    }
+
+    //CREATE  NEW CHANNEL
+    @PostMapping("channels")
+    public void add(@RequestBody ChannelDto  dto){
+        channelService.addChannel(ChannelDto.toEntity(dto));
     }
 
     //GET CHANNEL BY ID
     @GetMapping("channels/{id}")
-    public ResponseEntity<Channel> getChannelById(@PathVariable Long id) {
+    public Object getChannelById(@PathVariable Long id) {
         Optional<Channel> channel = channelService.getChannelById(id);
-        return channel.isPresent()? ResponseEntity.ok(channel.get()) : ResponseEntity.notFound().build();
-    }
-
-    //POST NEW CHANNEL
-    @PostMapping("channels")
-    public ResponseEntity<Channel> addChannel(@RequestBody Channel channel) {
-        Channel addedChannel = channelService.addChannel(channel);
-        return ResponseEntity.status(HttpStatus.CREATED).body(addedChannel);
+        ChannelDto channelDto = ChannelDto.fromEntity(channel);
+        if(channelDto == null) {
+            return  ResponseEntity.notFound().build();
+        } else {
+            return channelDto;
+        }
     }
 
     //DELETE CHANNEL
     @DeleteMapping("channels/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT) // 204
     public ResponseEntity<Void> deleteChannel(@PathVariable Long id) {
-        channelService.deleteChannel(id);
-        return ResponseEntity.noContent().build();
+        if(channelService.getChannelById(id).isEmpty() == true)
+            return ResponseEntity.notFound().build();
+        else {
+            channelService.deleteChannel(id);
+            return ResponseEntity.ok().build();
+        }
     }
 
     // UPDATE CHANNEL
     @PutMapping("channels")
-    public ResponseEntity update(@RequestBody Channel channel) {
-
-        Channel channelUpdate =  channelService.updateChannel(channel);
+    public ResponseEntity update(@RequestBody ChannelDto channelDto) {
+        Channel channelUpdate = channelService.updateChannel(ChannelDto.toEntity(channelDto));
         if(channelUpdate !=null)
-            return ResponseEntity.status(204).body("Channel successfully updated");
+            return ResponseEntity.status(200).build();
         else
-             return ResponseEntity.status(403).build();
+            return ResponseEntity.status(403).build();
 
     }
 }
